az oldal elhelyezése a profilmódosítás menüpontba és a dbfactory fgvények megírása hogy a backendel kommunikáljon

+++ API/server.js
@@ -369,7 +369,7 @@ app.post('/userupdate',(req,res)=>{
 app.post('/userdelete',(req,res)=>{
   if(session.Rights=="admin")
   {
-    dbPool.query(`DELETE FROM felhasznalok WHERE ID=${req.body.ID}`,(err,results)=>{
+    dbPool.query(`DELETE FROM felhasznalok WHERE ID=${req.body.ID} AND Jog NOT LIKE 'admin'`,(err,results)=>{
       if(err)throw err;
       res.json({message:"ok"});
     })
@@ -809,7 +809,7 @@ app.post('/opendelete',(req,res)=>{
   //helyfoglalás select
   
   app.post('/reservationSelect',(req,res)=>{
-    if(session.Rights=="user")
+    if(session.Rights=="user"||session.Rights=="admin")
     {
       let data={
         id:req.body.ID
@@ -850,7 +850,7 @@ app.post('/opendelete',(req,res)=>{
   //helyfoglalás delete
 
 app.post("/reservationDelete",(req,res)=>{
-  if(session.Rights=="etterem")
+  if(session.Rights=="user")
   {
     dbPool.query(`DELETE FROM helyfoglalas WHERE ID=${req.body.ID}`,(err,results)=>{
       if(err)throw err;
